adapt to new python3 behavior

+++ modules/photos_support.py
@@ -487,7 +487,7 @@ def save_member_face(params):
         face_photo_url = save_profile_photo(face, is_article=False)
     else:
         face_photo_url = None
-    if params.old_member_id >= 0:
+    if params.old_member_id and params.old_member_id > 0:
         q = (db.TblMemberPhotos.Photo_id == face.photo_id) & \
             (db.TblMemberPhotos.Member_id == params.old_member_id)
     else:
@@ -507,7 +507,7 @@ def save_member_face(params):
         rec.update_record(**data)
     else:
         db.TblMemberPhotos.insert(**data)
-        if params.old_member_id >= 0 and params.old_member_id != face.member_id:
+        if params.old_member_id and params.old_member_id > 0 and params.old_member_id != face.member_id:
             db(q).delete()
     member_name = member_display_name(member_id=face.member_id)
     db(db.TblPhotos.id==face.photo_id).update(Recognized=True)
@@ -682,6 +682,7 @@ def find_similar_photos(photo_list=None, time_budget=60):
     result = db((db.TblPhotos.id.belongs(all_dup_ids)) & (db.TblPhotos.deleted != True)).select()
     for photo_rec in result:
         photo_rec.dup_group = dic[photo_rec.id]
+    result = [Storage(rec) for rec in result]
     ##result = sorted(result, cmp=lambda prec1, prec2: +1 if prec1.dup_group > prec2.dup_group else -1 if prec1.dup_group < prec2.dup_group else +1 if prec1.id < prec2.id else -1)
     result = multisort(result, (('dup_group', False), ('id', True)))
     return result, candidates
